Renamed a couple helper files Parallelized web3 calls when possible
